% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_gpei_cases}
\alias{add_gpei_cases}
\title{Add GPEI website reported cases (DRAFT)}
\usage{
add_gpei_cases(
  azcontainer = suppressMessages(get_azure_storage_connection()),
  proxy_data_loc = "/Data/proxy/polio_proxy_data.csv",
  polis_pos_loc
)
}
\arguments{
\item{azcontainer}{Azure validated container object.}

\item{proxy_data_loc}{str location of proxy_data on EDAV}

\item{polis_pos_loc}{str location of latest positives dataset generated from POLIS API data. This is located in
/Data/polis and is a file name starting with "positives_2001-01-01".}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to add manually extracted GPEI cases to positives file and estimate points
based on lowest level admin data. This function is currently in draft form and isn't
intended to be used at this time.
}
\keyword{internal}
