% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_growth_rate_to_pop_data.R
\name{apply_growth_rate_to_pop_data}
\alias{apply_growth_rate_to_pop_data}
\title{Create estimated population counts for the new year}
\usage{
apply_growth_rate_to_pop_data(
  pop_dir = "GID/PEB/SIR/Data/pop",
  ctry_pop_path = file.path(pop_dir, "ctry.pop.long"),
  prov_pop_path = file.path(pop_dir, "prov.pop.long"),
  dist_pop_path = file.path(pop_dir, "dist.pop.long"),
  growth_path = file.path(pop_dir,
    "pop raw/WPP2024_GEN_F01_DEMOGRAPHIC_INDICATORS_COMPACT.xlsx"),
  edav = TRUE
)
}
\arguments{
\item{pop_dir}{\code{str} Path to the population directory.}

\item{ctry_pop_path}{\code{str} Regex to obtain the name of the country population file.}

\item{prov_pop_path}{\code{str} Regex to obtain the name of the province population file.}

\item{dist_pop_path}{\code{str} Regex to obtain the name of the district population file.}

\item{growth_path}{\code{str} Path to the growth rate file. The source of this dataset is in
the \href{https://population.un.org/wpp/downloads?folder=Standard\%20Projections&group=Most\%20used}{World Population Prospects}
website, using the "Compact" file.}

\item{edav}{\code{logical} Whether to read/write from EDAV. Defaults to \code{TRUE}.}
}
\value{
\code{list} List containing the population data, including the new year, invisibly.
}
\description{
Using growth rate estimates for the latest year available, the function
estimates the population counts of the new year. This is typically done when the
new population estimates for the new year is unavailable. New official population
counts are typically reported in April through August of the new year.
}
\examples{
\dontrun{
apply_growth_rate_to_pop_data
}
}
