% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{archive_log}
\alias{archive_log}
\title{function to archive log report}
\usage{
archive_log(
  log_file = Sys.getenv("POLIS_LOG_FILE"),
  polis_data_folder = Sys.getenv("POLIS_DATA_CACHE")
)
}
\arguments{
\item{log_file}{str: location of POLIS log file}

\item{polis_data_folder}{str: location of the POLIS data folder}
}
\description{
Function to read in log file and archive entries older than 3 months
}
