% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_and_reconcile}
\alias{bind_and_reconcile}
\title{Reconcile classes and bind two tibbles}
\usage{
bind_and_reconcile(new_data, old_data)
}
\arguments{
\item{new_data}{tibble: Tibble to be converted and bound}

\item{old_data}{tibble: Tibble to be referenced}
}
\value{
tibble: bound tibble
}
\description{
Reconcile classes and bind two tibbles
}
