% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_urls.R
\name{call_urls}
\alias{call_urls}
\title{Call multiple URLs}
\usage{
call_urls(urls)
}
\arguments{
\item{urls}{array of url strings}
}
\value{
tibble with all data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Call multiple URLs
}
\keyword{internal}
