% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_static_files}
\alias{check_missing_static_files}
\title{Check for missing static files}
\usage{
check_missing_static_files(
  core_files_folder_path,
  edav_flag = Sys.getenv("POLIS_EDAV_FLAG")
)
}
\arguments{
\item{core_files_folder_path}{`str` Path to the `core_files_to_combine` folder.}

\item{edav}{`bool` Check static files on EDAV?}
}
\value{
`list` A list of missing files.
}
\description{
Performs a check for the static files.
}
\keyword{internal}
