% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_static_files}
\alias{check_static_files}
\title{Check the historical static files}
\usage{
check_static_files(
  polis_data_folder,
  edav_flag = Sys.getenv("POLIS_EDAV_FLAG")
)
}
\arguments{
\item{polis_data_folder}{`str` Path to the POLIS data folder.}

\item{edav}{`bool` Check static files on EDAV?}
}
\value{
`list` A list of missing files.
}
\description{
Performs a check for the static files.
}
\keyword{internal}
