% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_case_data}
\alias{clean_case_data}
\title{Preprocess population data into flat files}
\usage{
clean_case_data(
  case_path,
  crosswalk_data,
  edav = Sys.getenv("POLIS_EDAV_FLAG")
)
}
\arguments{
\item{case_path}{`str` Path to case data.}

\item{crosswalk_data}{`tibble` The crosswalk table.}

\item{edav}{`bool` Is the case data located on EDAV? Defaults to the
EDAV flag during [init_tidypolis()].}
}
\value{
`tibble` cleaned Case data.
}
\description{
The function performs cleaning of the Case table. In particular,
de-duplication, renaming of variables via crosswalk, and removal of
empty columns.
}
\details{
Process POLIS population data using CDC and other standards
 readr dplyr
str: "cdc" or "who" (default)
 tibble: WHO POLIS population file, defaults to tidypolis folder
 list with tibble for ctry, prov and dist
process_pop <- function(type = "who", pop_file = readr::read_rds(file.path(Sys.getenv("POLIS_DATA_FOLDER"), "data", "pop.rds"))){

  subset to <= 15
  pop_file <- pop_file |>
    filter(AgeGroupName == "0 to 15 years")

  #extract into country prov and dist

  x <- lapply(unique(pop_file$Admin0Name), function(x){
    pop_file |>
      filter(is.na(Admin1Name) & is.na(Admin2Name)) |>
      rename(year = Year, u15pop = Value, GUID = Admin0GUID, ctry = Admin0Name) |>
      mutate(u15pop = as.integer(u15pop)) |>
      arrange(year) |>
      filter(ctry == x) |>
      group_by(year) |>
      filter(!is.na(u15pop)) |>
      filter(UpdatedDate == max(UpdatedDate, na.rm = T)) |>
      ungroup() |>
      select(ctry, year, u15pop, GUID) |>
      full_join(tibble(ctry = x, year = 2000:(lubridate::year(Sys.time()))), by = c("ctry", "year"))
  }) |>
    bind_rows()

}

Clean Case data
}
\keyword{internal}
