% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_case_table}
\alias{clean_case_table}
\title{Clean Case table}
\usage{
clean_case_table(path, crosswalk_data, edav = Sys.getenv("POLIS_EDAV_FLAG"))
}
\arguments{
\item{path}{\code{str} File path to the table.}

\item{edav}{\code{bool} Is the case data located on EDAV? Defaults to the
EDAV flag.}

\item{crosswalk}{\code{tibble} The crosswalk table. This is the output of
\code{\link[=get_crosswalk_data]{get_crosswalk_data()}}.}
}
\value{
\code{tibble} cleaned Case data.
}
\description{
The function performs cleaning of the Case table. In particular,
de-duplication, renaming of variables via crosswalk, and removal of
empty columns.
}
\keyword{internal}
