% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_virus_table}
\alias{clean_virus_table}
\title{Clean the virus table}
\usage{
clean_virus_table(path, crosswalk, edav = Sys.getenv("POLIS_EDAV_FLAG"))
}
\arguments{
\item{path}{\code{str} File path to the table.}

\item{crosswalk}{\code{tibble} The crosswalk table. This is the output of
\code{\link[=get_crosswalk_data]{get_crosswalk_data()}}.}

\item{edav}{\code{bool} Data located on EDAV? Defaults to the
EDAV flag.}
}
\value{
\code{tibble} Cleaned virus table.
}
\description{
Clean the virus table
}
\keyword{internal}
