% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_dates}
\alias{cluster_dates}
\title{SIA Cluster Function}
\usage{
cluster_dates(x, seed = 1234, method = "kmeans", grouping_days = 365)
}
\arguments{
\item{x}{df: data to be clustered}

\item{seed}{num}

\item{method}{str cluster method to use, can be "kmeans" or "mindate"}

\item{grouping_days}{int:}
}
\description{
this function identifies "cluster" or OBX response so we can identify rounds
}
