% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_dates_for_sias}
\alias{cluster_dates_for_sias}
\title{Manager function for running cluster_dates() funciton}
\usage{
cluster_dates_for_sias(sia)
}
\arguments{
\item{sia}{dataframe of SIAs to identify rounds by vaccine type}
}
\description{
manager function to run the cluster_dates() function using helper function run_cluster_dates to cluster SIAs by type
}
