% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{collate_file_extracts}
\alias{collate_file_extracts}
\title{Collate extracts into one table}
\usage{
collate_file_extracts(table_data)
}
\value{
\code{conn} A parquet connection.
}
\description{
This function collates the extracts into one table.
}
\keyword{internal}
