% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_core_ready_dir}
\alias{create_core_ready_dir}
\title{Helper function to create Core Ready Archive and Change Log}
\usage{
create_core_ready_dir(polis_data_folder, timestamp)
}
\arguments{
\item{polis_data_folder}{\code{str} Location of the POLIS data folder.}

\item{timestamp}{\code{str} Folder name as the timestamp.}
}
\description{
Helper function to create Core Ready Archive and Change Log
}
\keyword{internal}
