% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cred_file}
\alias{create_cred_file}
\title{Set up local credentials file}
\usage{
create_cred_file(polis_data_folder)
}
\arguments{
\item{polis_data_folder}{str: location of POLIS data folder}
}
\value{
boolean for folder creation
}
\description{
Create creds file
}
\keyword{internal}
