% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{create_extract_file}
\alias{create_extract_file}
\title{Create an extract file of an API table}
\usage{
create_extract_file(table_data, extract)
}
\arguments{
\item{table_data}{\code{tibble} One row tibble with the data for a specific table.}

\item{extract}{\code{tibble} Newly updated data for the table specified.}
}
\value{
\code{NULL} upon success
}
\description{
Create an extract file of an API table and put it in the extract
folder of the API table. The extract only contains new data that were updated
since the last API pull.
}
\keyword{internal}
