% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_frontend_case_table.R
\name{crosswalk_frontend_case_table}
\alias{crosswalk_frontend_case_table}
\title{Transform front-end case table names to backend case table equivalent}
\usage{
crosswalk_frontend_case_table(
  case_table_path,
  output_dir_path = dirname(case_table_path),
  output_type = "rds"
)
}
\arguments{
\item{case_table_path}{\code{str} Path to the front-end case table. Must be a csv file.}

\item{output_dir_path}{\code{str} Path to the directory the case table should be outputted to.
Defaults to the directory of the front-end case table.}

\item{output_type}{\code{str} The format the file should be outputted as. Valid outputs are:
rds, rda, csv, parquet. Defaults to rds.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Transforms the front-end case table columns to the equivalent column of the
backend. Note, the front-end case table contains fewer columns compared to the backend.
The outputted file can be written as an RDS, csv, parquet, or qs2.
}
\examples{
\dontrun{
case_table_path <- "C:/Users/ABC1/Downloads/front_end_case_table.csv"
crosswalk_frontend_case_table(case_table_path)
}
}
