% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.compare.dataframe.cols}
\alias{f.compare.dataframe.cols}
\title{Check if sets of columns are the same and stop if not}
\usage{
f.compare.dataframe.cols(old, new)
}
\arguments{
\item{old}{str/array: array of character column names}

\item{new}{str/array: array of character colunm names}
}
\value{
str: Stop error or continue
}
\description{
Check if sets of names are the same and return stop error if not
}
