% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.compare.metadata}
\alias{f.compare.metadata}
\title{Compare meta data outputs between two datasets}
\usage{
f.compare.metadata(new_table_metadata, old_table_metadata)
}
\arguments{
\item{new_table_metadata}{tibble}

\item{old_table_metadata}{tibble}
}
\value{
meta data comparisons
}
\description{
compare meta data outputs between two datasets
}
