% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.datecheck.onset}
\alias{f.datecheck.onset}
\title{Function for data qa check in AFP line list cleaning}
\usage{
f.datecheck.onset(date1, date2)
}
\arguments{
\item{date1}{date 1 is the date to be checked against date2}

\item{date2}{date 2 is the date of onset such that the date should be after onset}
}
\value{
quality controlled date variable
}
\description{
function creates a new variable when combined with a mutate statement in R code
}
\keyword{internal}
