% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.download.compare.01}
\alias{f.download.compare.01}
\title{Compare downloaded data}
\usage{
f.download.compare.01(old.download, new.download)
}
\arguments{
\item{old.download}{tibble}

\item{new.download}{tibble}
}
\value{
tibble: variables that are new or unaccounted for
}
\description{
Compared downloade data
}
