% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.pre.stsample.01}
\alias{f.pre.stsample.01}
\title{Sample points for missing lat/lon}
\usage{
f.pre.stsample.01(df01, global.dist.01)
}
\arguments{
\item{df01}{tibble: table of afp data}

\item{global.dist.01}{sf: spatial file of all locations}
}
\value{
tibble with lat/lon for all unsampled locations
}
\description{
Create random samples of points for missing GPS data
}
