% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.summarise.metadata}
\alias{f.summarise.metadata}
\title{Summarize metadata from a tibble}
\usage{
f.summarise.metadata(dataframe, categorical_max = 10)
}
\arguments{
\item{dataframe}{tibble}

\item{categorical_max}{int: maximum number of categories considered}
}
\value{
tibble: metadata
}
\description{
Summarize metadata from tibble
}
\keyword{internal}
