% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_current_week.R
\name{filter_current_week}
\alias{filter_current_week}
\title{Filters the cases based on the current epi-week as defined by WHO}
\usage{
filter_current_week(tibble, date_col, week_floor_date)
}
\arguments{
\item{tibble}{\code{tibble} A tibble containing the virus table.}

\item{date_col}{\code{str} The date column to filter on.}

\item{week_floor_date}{\code{Date} The date to anchor on, typically the Monday of the
current week.}
}
\value{
\code{tibble} Filtered virus table.
}
\description{
Filters the cases based on the current epi-week as defined by WHO
}
\keyword{internal}
