% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_latest_archive}
\alias{find_latest_archive}
\title{Find Latest Archive in Core Ready Files}
\usage{
find_latest_archive(polis_data_folder)
}
\arguments{
\item{polis_data_folder}{Character string specifying the path to the main
data folder containing Core Ready Files}
}
\value{
Character string containing either the name of the most recent
archive folder or the current timestamp if no archives exist
}
\description{
This function searches for and returns the name of the most recent archive
folder in the Core Ready Files directory. If no archive is found, it returns
the current timestamp based on prespecfied time format.
}
\details{
The function performs the following steps:
\enumerate{
\item Lists contents of the Archive folder in Core Ready Files
\item Converts folder names to datetime objects
\item If archives exist, returns name of the most recent one
\item If no archives exist, returns current timestamp
}
}
\examples{
\dontrun{
latest_archive <- find_latest_archive("path/to/polis/data")
}
}
\keyword{internal}
