% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{force_compile_raw_extract}
\alias{force_compile_raw_extract}
\title{Force compile an extract folder}
\usage{
force_compile_raw_extract(table_data)
}
\arguments{
\item{table_data}{\code{tibble} A one row tibble with table information}
}
\value{
\code{tibble} Updated table cache, invisibly.
}
\description{
Force compile an extract folder
}
\examples{
\dontrun{

init_tidypolis(edav = T)
cache <- get_polis_cache()
force_compile_raw_extract(cache[2, ]) # this corresponds to the virus table_data
}
}
