% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_api_virus_table.R
\name{format_api_virus_table}
\alias{format_api_virus_table}
\title{Helper function to standardize names of the API virus table}
\usage{
format_api_virus_table(api_virus_table)
}
\arguments{
\item{api_virus_table}{\code{tibble} The virus table downloaded from the POLIS API.}
}
\value{
\code{tibble} A virus table with the same names as those found in the preprocessed positives file.
}
\description{
Allows comparison to the preprocessed positives file.
}
\keyword{internal}
