% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{freeze_polis_data}
\alias{freeze_polis_data}
\title{Build a freeze of POLIS data on a specific date}
\usage{
freeze_polis_data()
}
\value{
Location of zipped and frozen POLIS data
}
\description{
Zip all POLIS data that is used to work on a specific date
}
