% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_crosswalk_data}
\alias{get_crosswalk_data}
\title{Get crosswalk data}
\usage{
get_crosswalk_data(
  file_loc = file.path(Sys.get("DATA_FOLDER"), "misc/crosswalk.rds")
)
}
\arguments{
\item{file_loc}{str: location of crosswalk file}
}
\value{
tibble: crosswalk data
}
\description{
Get all data from crosswalk location
}
