% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_env_site_data}
\alias{get_env_site_data}
\title{Function to get cached env site data}
\usage{
get_env_site_data(
  file_loc = file.path(Sys.getenv("POLIS_DATA_FOLDER"), "misc", "env_sites.rds")
)
}
\arguments{
\item{file_loc}{\code{path} Path to the env_sites.rds file.}
}
\value{
\code{tibble} Env site list
}
\description{
Function to get cached env site data
}
\keyword{internal}
