% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{get_full_table}
\alias{get_full_table}
\title{Should full table be downloaded?}
\usage{
get_full_table(id_error, table_data)
}
\arguments{
\item{id_error}{\code{logical} Whether the ID column is function.}

\item{table_data}{\code{tibble} One row tibble with the data for a specific table.}
}
\value{
\code{logical} Whether to download the full table or not.
}
\description{
Should full table be downloaded?
}
\keyword{internal}
