% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaning.R
\name{get_lab_locs}
\alias{get_lab_locs}
\title{Table of information regarding testing labs in each country}
\usage{
get_lab_locs(lab_locs_path = NULL, use_edav = TRUE)
}
\arguments{
\item{lab_locs_path}{\code{str} Location of testing lab locations. Default is \code{NULL}. Will download from EDAV, if use_edav=TRUE.}

\item{use_edav}{\code{logical} Whether to obtain lab locs data from EDAV Defaults to \code{TRUE}.}
}
\value{
\code{tibble} A table containing the test lab location information.
}
\description{
Imports information on testing labs for each country, either from a CSV file
or downloaded from EDAV. If no argument is passed, the function will download
the table from EDAV.
}
\examples{
\dontrun{
ctry.seq <- get_lab_locs()
}

}
