% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_growth_rate_to_pop_data.R
\name{get_new_year_pop}
\alias{get_new_year_pop}
\title{Calculate pop counts for new year}
\usage{
get_new_year_pop(pop_w_gr)
}
\arguments{
\item{pop_w_gr}{\code{tibble} Population file with growth rates.}
}
\value{
\code{tibble} New year population data.
}
\description{
Applies the growth rate from the previous year to calculate new populations
for the new year.
}
\keyword{internal}
