% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_polis_cache}
\alias{get_polis_cache}
\title{Load local POLIS cache}
\usage{
get_polis_cache(cache_file = Sys.getenv("POLIS_CACHE_FILE"), .table = NULL)
}
\arguments{
\item{cache_file}{\code{str} Location of cache file.}

\item{.table}{\code{str} Table(s) to be loaded. A vector of table names can be passed. Defaults to \code{NULL},
which loads the cache file in full.}
}
\value{
\code{tibble} Metadata of the POLIS API table specified.
}
\description{
Pull metadata for a particular POLIS table.
}
\examples{
\dontrun{
cache <- get_polis_cache()
}

}
