% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_ids.R
\name{get_table_ids}
\alias{get_table_ids}
\title{Get the unique IDs for a particular POLIS table}
\usage{
get_table_ids(table_data, api_key = Sys.getenv("POLIS_API_KEY"))
}
\arguments{
\item{table_data}{\code{tibble} One row tibble with the associated table data.}

\item{api_key}{\code{str} POLIS API Key.}
}
\value{
\code{str} A vector of IDs.
}
\description{
Return the full IDs available in API table.
}
