% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_tidypolis}
\alias{init_tidypolis}
\title{A function to intitialize a POLIS data folder}
\usage{
init_tidypolis(polis_folder = "POLIS", edav = TRUE, api_debug = FALSE)
}
\arguments{
\item{polis_folder}{\code{str} Location of folder where to store all information from POLIS.}

\item{edav}{\code{logical} Should the system use EDAV as it's cache; default \code{TRUE}.}

\item{api_debug}{boolean: if true will log all api calls}
}
\value{
Messages on process.
}
\description{
Initialize API Key and local data cache for tidypolis. Inspired by the
tidycensus process for managing their API secrets.
}
\examples{
\dontrun{
# create a polis data cache on CDC EDAV
init_tidypolis(polis_folder = "POLIS", data_folder = "Data", edav = T)
init_tidypolis("C:/Users/user_name/Desktop/polis",
  "C:/Users/user_name/Desktop/data",
  edav = F
)
# create a polis data cache on your local desktop
}
}
