% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_growth_rate_to_pop_data.R
\name{load_growth_rates}
\alias{load_growth_rates}
\title{Load population growth rates}
\usage{
load_growth_rates(growth_path, edav)
}
\arguments{
\item{growth_path}{\code{str} Path to WPP Excel file.}
}
\value{
\code{tibble} Growth rates for each country by year.
}
\description{
Load population growth rates
}
\details{
Loads and formats the population growth rate file for use in R.

The Excel file uses the "Estimates" tab. The first few rows (16 rows or so) are deleted for ease of loading into
R. The source of this dataset is in the \href{https://population.un.org/wpp/downloads?folder=Standard\%20Projections&group=Most\%20used}{World Population Prospects}
website, using the "Compact" file.
}
\keyword{internal}
