% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaning.R
\name{load_lab_data}
\alias{load_lab_data}
\title{Function to load the raw lab data locally}
\usage{
load_lab_data(lab_data_path, sheet_name = NULL)
}
\arguments{
\item{lab_data_path}{\code{str} File path as a string to the lab data.}

\item{sheet_name}{\code{str} Name of the sheet to load. This is optional in cases
of an Excel sheet with multiple tabs.}
}
\value{
\code{tibble} Lab data loaded from the CSV or Excel file path.
}
\description{
This a function to load lab data that are either CSVs or Excel files.
}
\examples{
\dontrun{
lab_data_path <- "C:/Users/ABC1/Desktop/lab_data.csv"
lab_data <- load_lab_data(lab_data_path)
}

}
