% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_report}
\alias{log_report}
\title{Function to read and report on latest log file entries}
\usage{
log_report(
  log_file = Sys.getenv("POLIS_LOG_FILE"),
  polis_data_folder = Sys.getenv("POLIS_DATA_CACHE")
)
}
\arguments{
\item{log_file}{\code{str} location of POLIS log file}

\item{polis_data_folder}{\code{str} location of the POLIS data folder}
}
\description{
Read log entries from the latest download and preprocessing run,
create report to send to Teams chat. This function is not meant to be used outside
of the SIR team at its current iteration. Future versions may allow for
more flexibility in the function.
}
