% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaning.R
\name{parse_lab_date}
\alias{parse_lab_date}
\title{Robust data parser for lab dates from Excel file}
\usage{
parse_lab_date(x)
}
\description{
Robust data parser for lab dates from Excel file
}
\keyword{internal}
