% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaning.R
\name{prep_lab_data}
\alias{prep_lab_data}
\title{Prep lab data}
\usage{
prep_lab_data(
  lab_data_path,
  afp_data = NULL,
  lab_locs_path = NULL,
  use_edav = TRUE,
  save_to_edav = FALSE,
  output_lab_checks = FALSE,
  all_regions = FALSE
)
}
\arguments{
\item{lab_data_path}{\code{string} filepath to lab datasets.}

\item{afp_data}{\code{tibble} AFP linelist. If using tidypolis, use \code{raw_data$afp}.}

\item{lab_locs_path}{\code{str} Location of testing lab locations. Default is \code{NULL}. Will download from EDAV, if use_edav=TRUE.}

\item{use_edav}{\code{logical} Whether to obtain lab locs data from EDAV Defaults to \code{TRUE}.}

\item{save_to_edav}{\code{logical} Whether to save Rda file to EDAV, Defaults to \code{FALSE}.}

\item{output_lab_checks}{\code{logical} Whether to save CSV file with cases with implausible dates to the working directory, Defaults to \code{FALSE}.}

\item{all_regions}{\code{logical} Whether to read one combined lab extract named \code{All_regions Lab Extract}. Defaults to \code{FALSE}.}
}
\value{
\code{tibble} Cleaned lab data.
}
\description{
Main lab data preparation function.
Requires Excel files from WHO saved to a local file path.
Expects the names of Excel files to contain the pattern \verb{[Region] Lab Extract}
and the Sheets within the Excel files to contain the pattern \verb{[Region] Lab data}
If \code{all_regions = TRUE}, expects one combined extract named \code{All_regions Lab Extract}.
}
\details{
uploads to EDAV if use_edav = TRUE, otherwise, returns the lab data frame as an output that can be written as a CSV file to a local directory

Will also output a CSV file with cases with implausible dates to the working directory if output_lab_checks=TRUE
}
\examples{
\dontrun{
With lab data and lab locs stored locally:
raw_data<-sirfunctions::get_all_polio_data(attach.spatial.data = FALSE)
lab_data <- prep_lab_data(
lab_data_path = "C:/Users/abc1/labdata/2026-06-30 Lab Extract (AFP only since 2023-30Jun2026)/",
afp_data = raw_data$afp,
lab_locs_path = "C:/Users/abc1/labdata/Routine lab testing locations.csv"
use_edav=FALSE,
)

# Using EDAV for lab locs:
raw_data<-sirfunctions::get_all_polio_data(attach.spatial.data = FALSE)
lab_data <- prep_lab_data(
lab_data_path = "C:/Users/abc1/labdata/2026-06-30 Lab Extract (AFP only since 2023-30Jun2026)/",
afp_data = raw_data$afp,
lab_locs_path = NULL,
use_edav=TRUE
)
}
}
