% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data retrieved from POLIS}
\usage{
preprocess_data(type = "cdc", who_region = NULL, output_format = "parquet")
}
\arguments{
\item{type}{\code{str} Specify the type of preprocessing to complete}

\item{who_region}{\code{str} Optional WHO region to filter data
Available inputs include AFRO, AMRO, EMRO, EURO, SEARO and  WPRO.}

\item{output_format}{\code{str} Output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'parquet'.}
}
\value{
Analytic files.
}
\description{
Create standard analytic datasets from raw POLIS data
}
\examples{
\dontrun{
preprocess_data(type = "cdc") # must run init_tidypolis to specify POLIS data location first
}
}
