% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spatial.R
\name{process_spatial}
\alias{process_spatial}
\title{A function to process WHO spatial datasets}
\usage{
process_spatial(
  gdb_folder = "GID/PEB/SIR/Data/spatial/WHO_POLIO_GLOBAL_GEODATABASE.gdb",
  output_folder = "GID/PEB/SIR/Data/spatial",
  edav = TRUE,
  azcontainer = suppressMessages(sirfunctions::get_azure_storage_connection())
)
}
\arguments{
\item{gdb_folder}{\code{str} The folder location of spatial datasets, should end with .gdb,
if on edav the gdb will need to be zipped, ensure that the gdb and the zipped file name are the same.}

\item{output_folder}{\code{str} Folder location to write outputs to.}

\item{edav}{\code{logical} Whether gdb is on EDAV or local.}

\item{azcontainer}{\verb{Azure container} Azure storage container.}
}
\description{
a function to process WHO spatial datasets
}
\examples{
\dontrun{
process_spatial(
  gdb_folder = "local_path/GEODATABASE.gdb",
  output_folder = "local_path",
  edav = F
)
process_spatial(
  gdb_folder = "edav_path/GEODATABASE.gdb",
  output_folder = "edav_path",
  edav = T
)
}
}
