% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Remove empty columns}
\usage{
remove_empty_columns(dataframe)
}
\arguments{
\item{dataframe}{\code{tibble} df}
}
\value{
\code{tibble} without any empty columns
}
\description{
Utility function to return tibble without empty columns
}
\keyword{internal}
