% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_via_crosswalk}
\alias{rename_via_crosswalk}
\title{Rename variables via crosswalk}
\usage{
rename_via_crosswalk(api_data, crosswalk, table_name)
}
\arguments{
\item{api_data}{`tibble` Data pulled from the POLIS API.}

\item{crosswalk}{`tibble` Crosswalk file. The crosswalk file can be imported
as a tibble by using [get_crosswalk_data()].}

\item{table_name}{`str` Name of table to be crosswalked. The following values
are valid: `"Case", "EnvSample", "SubActivity", "Virus", "Activity"`.}
}
\value{
`tibble` Renamed `api_data` using the crosswalk file.
}
\description{
Rename variables in tibble using crosswalk data. The crosswalk
file is created by the PEB SIR Team at CDC. Please request the crosswalk
data before running pre-processing.
}
\keyword{internal}
