% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{request_input}
\alias{request_input}
\title{Modified readlines function to simplify user interface}
\usage{
request_input(request, vals, error_resp = "Valid input not chosen", max_i = 3)
}
\arguments{
\item{request}{str:}

\item{vals}{array:}

\item{error_resp}{str:}

\item{max_i}{int:}
}
\value{
val from \code{vals} or stops output
}
\description{
Manage input
}
\keyword{internal}
