% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_cluster_dates}
\alias{run_cluster_dates}
\title{Wrapper function for cluster_dates, includes error checking and cluster cache management}
\usage{
run_cluster_dates(
  data,
  cache_folder = file.path(Sys.getenv("DATA_FOLDER"), "sia_cluster_cache"),
  min_obs = 4,
  type
)
}
\arguments{
\item{data}{df dataframe on which to run cluster dates function}

\item{cache_folder}{str location of sia cluster cache on CDC EDAV}

\item{min_obs}{int}

\item{type}{str vaccine type}
}
\description{
Wrapper around the cluster_dates function to do some error checking
}
