% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{run_diagnostics}
\alias{run_diagnostics}
\title{Run diagnostic test on polis connections}
\usage{
run_diagnostics()
}
\value{
tibble with diagnostic results
}
\description{
Run diagnostics of API connection with POLIS
}
