% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_table_diagnostic.R
\name{run_single_table_diagnostic}
\alias{run_single_table_diagnostic}
\title{Run single table diagnostic}
\usage{
run_single_table_diagnostic(.table, key = Sys.getenv("POLIS_API_Key"))
}
\arguments{
\item{.table}{\code{str} Table name. Valid values are "virus", "case", "human_specimen",
"environmental_sample", "activity", "sub_activity", "lqas", "im",  "population".}

\item{key}{\code{str} POLIS API Key.}
}
\value{
\code{tibble} Diagnostic data on whether data was successfully pulled from POLIS.
}
\description{
Run diagnostics on a single API table. These are
}
