% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_archive_old_files}
\alias{s1_archive_old_files}
\title{Archives the old files in the data folder}
\usage{
s1_archive_old_files(
  polis_data_folder,
  timestamp,
  output_folder_name,
  output_format
)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder}

\item{timestamp}{\code{str} Time stamp folder name}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{output_format}{\code{str} output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}
}
\description{
Archives the old files in the data folder
}
\keyword{internal}
