% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_clean_subactivity_table}
\alias{s1_clean_subactivity_table}
\title{Clean the subactivity table}
\usage{
s1_clean_subactivity_table(
  path,
  activity_table,
  crosswalk,
  long_dist_sf,
  edav = Sys.getenv("POLIS_EDAV_FLAG")
)
}
\arguments{
\item{path}{\code{str} File path to the table.}

\item{activity_table}{\code{tibble} Output of \code{\link[=clean_activity_table]{clean_activity_table()}}.}

\item{crosswalk}{\code{tibble} The crosswalk table. This is the output of
\code{\link[=get_crosswalk_data]{get_crosswalk_data()}}.}

\item{long_dist_sf}{\code{tibble} District shapefile in long format. The output of
\code{sirfunctions::load_clean_dist_sp(type = "long")}.}

\item{edav}{\code{logical} Data located on EDAV? Defaults to the
EDAV flag.}
}
\value{
\code{tibble} Cleaned Sub-activity table.
}
\description{
The function performs cleaning of the sub-activity table. In particular,
de-duplication, renaming of variables via crosswalk, and removal of
empty columns.
}
\keyword{internal}
