% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_create_change_log}
\alias{s1_create_change_log}
\title{Create the change log for table data}
\usage{
s1_create_change_log(
  polis_data_folder,
  file,
  timestamp,
  api_case_data,
  api_es_data,
  api_virus_data,
  api_subactivity_data
)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder}

\item{file}{\code{str} Name of the file}

\item{timestamp}{\code{str} Name of the timestamp folder}

\item{api_case_data}{\code{tibble} Cleaned case data}

\item{api_es_data}{\code{tibble} Cleaned ES data}

\item{api_virus_data}{\code{tibble} Cleaned virus data}

\item{api_subactivity_data}{\code{tibble} Cleaned subactivity data}
}
\value{
\code{NULL}, if successful
}
\description{
Create the change log for table data
}
\keyword{internal}
