% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_get_most_recent_files}
\alias{s1_get_most_recent_files}
\title{Gets the most recent files in the Core Ready folder}
\usage{
s1_get_most_recent_files(polis_data_folder, patterns)
}
\arguments{
\item{polis_data_folder}{\code{str} Location of the POLIS data folder.}

\item{patterns}{\code{list} A list of patterns to select files.}
}
\value{
\code{list} With names of the most recent files that matches the patterns.
}
\description{
Gets the most recent files in the Core Ready folder
}
\keyword{internal}
