% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s2_check_missingness}
\alias{s2_check_missingness}
\title{Check missingness in key AFP variables}
\usage{
s2_check_missingness(data, type = "AFP", polis_data_folder)
}
\arguments{
\item{data}{A data frame containing AFP surveillance data}

\item{type}{String indicating the type of data (e.g., "AFP")}

\item{polis_data_folder}{String path to the POLIS data folder}
}
\value{
Invisibly returns the missingness summary
}
\description{
Analyzes and exports information about missingness in key variables that are
critical for AFP surveillance. This helps with data quality assessment.
}
