% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s2_fully_process_afp_data}
\alias{s2_fully_process_afp_data}
\title{Process AFP Data Pipeline}
\usage{
s2_fully_process_afp_data(
  polis_data_folder,
  polis_folder,
  long.global.dist.01,
  timestamp,
  latest_folder_in_archive_path,
  output_folder_name,
  output_format,
  archive = TRUE
)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder containing
Core_Ready_Files.}

\item{polis_folder}{\code{str} Path to the main POLIS folder.}

\item{long.global.dist.01}{\code{sf} Global district lookup table for GUID
validation.}

\item{timestamp}{\code{str} Time stamp}

\item{latest_folder_in_archive_path}{\code{str} The path to the latest archive folder.}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}

\item{archive}{Logical. Whether to archive previous output directories
before overwriting. Default is \code{TRUE}.}
}
\description{
This function processes AFP data through multiple standardization and
validation steps, including checking for duplicates, standardizing dates,
classifying cases, and processing coordinates.
}
