% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s2_process_coordinates}
\alias{s2_process_coordinates}
\title{Process GPS coordinates for AFP cases}
\usage{
s2_process_coordinates(data, polis_data_folder, polis_folder)
}
\arguments{
\item{data}{\code{tibble} A data frame containing AFP data with admin GUIDs}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder containing
Core_Ready_Files.}

\item{polis_folder}{\code{str} Path to the main POLIS folder.}
}
\value{
A tibble with processed coordinate data
}
\description{
Handles missing coordinates, validates existing ones, and ensures that all
AFP cases have accurate geographic coordinates for mapping and spatial
analysis.
}
