% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_check_duplicates}
\alias{s3_sia_check_duplicates}
\title{Check SIA dataset for duplicates}
\usage{
s3_sia_check_duplicates(sia.05)
}
\arguments{
\item{sia.05}{\code{tibble} The latest SIA download with variables checked and
against the last download, CDC variables created and GUIDs validated}
}
\value{
\code{tibble} sia.06 SIA data with CDC variables enforced, GUIDs validated
and duplicates removed
}
\description{
Check SIA dataset for duplicates
}
\keyword{internal}
