% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_check_guids}
\alias{s3_sia_check_guids}
\title{Check SIA GUIDS to ensure they match the standard spatial dataset}
\usage{
s3_sia_check_guids(sia.02, long.global.dist.01)
}
\arguments{
\item{sia.02}{\code{tibble} The latest SIA download with variables checked and
against the last download and CDC variables created}

\item{long.global.dist.01}{\code{sf} a cleaned global spatial dataset of districts
in "long" format for years meaning a district spatial object that is valid
from 2000-2005 will be represented 6 times, once for each year it is valid}
}
\value{
\code{tibble} sia.03 SIA data with CDC variables enforced and GUIDs validated
}
\description{
Check SIA GUIDS to ensure they match the standard spatial dataset
}
\keyword{internal}
