% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_cluster_dates}
\alias{s3_sia_cluster_dates}
\title{Manager function for running cluster_dates() function}
\usage{
s3_sia_cluster_dates(sia)
}
\arguments{
\item{sia}{\code{tibble} tibble of SIAs to identify rounds by vaccine type}
}
\value{
\code{NULL} silently.
}
\description{
manager function to run the cluster_dates() function using helper function s3_run_cluster_dates to cluster SIAs by type
}
\keyword{internal}
