% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_cluster_dates_by_vax_type}
\alias{s3_sia_cluster_dates_by_vax_type}
\title{Wrapper function for cluster_dates, includes error checking and cluster cache management}
\usage{
s3_sia_cluster_dates_by_vax_type(
  data,
  cache_folder = file.path(Sys.getenv("POLIS_DATA_FOLDER"), "misc", "sia_cluster_cache"),
  min_obs = 4,
  type
)
}
\arguments{
\item{data}{\code{tibble} dataframe on which to run cluster dates function}

\item{cache_folder}{\code{str} location of sia cluster cache on CDC EDAV}

\item{min_obs}{\code{int} the minumum number of dates needed to consider clustering
using k-means versus using the minimum date available, defaults to 4 for
global Polio SIAs}

\item{type}{\code{str} vaccine type}
}
\description{
Wrapper around the cluster_dates function to do some error checking
}
