% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_combine_historical_data}
\alias{s3_sia_combine_historical_data}
\title{Read in SIA data pre 2020 and combine with current SIA data}
\usage{
s3_sia_combine_historical_data(sia.new)
}
\arguments{
\item{sia.06}{\code{tibble} The latest SIA download with variables checked and
against the last download, CDC variables created, GUIDs validated and
deduplicated}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}
}
\value{
\code{tibble} sia.clean.01 All historical SIA data
}
\description{
Read in SIA data pre 2020 and combine with current SIA data
}
\keyword{internal}
