% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_evaluate_unmatched_guids}
\alias{s3_sia_evaluate_unmatched_guids}
\title{Evaluate SIA data with unmatched spatial data}
\usage{
s3_sia_evaluate_unmatched_guids(sia.05, polis_data_folder)
}
\arguments{
\item{sia.05}{\code{tibble} the output of s3_sia_check_guids()}

\item{polis_data_folder}{The POLIS data folder.}
}
\description{
Looks at the output from the GUID matching for SIAs and evalutes
all the SIAs that did not have a matching GUID
}
\keyword{internal}
