% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_export_outputs}
\alias{s3_sia_export_outputs}
\title{Export sia.clean.01 as the clean dataset for SIA.}
\usage{
s3_sia_export_outputs(
  sia.clean.01,
  polis_folder = Sys.getenv("POLIS_DATA_FOLDER"),
  polis_data_folder = file.path(polis_folder, "data"),
  output_folder_name,
  output_format
)
}
\arguments{
\item{sia.clean.01}{\code{tibble} all cleaned and historical SIA data without rounds}

\item{polis_folder}{\code{str} Path to the POLIS folder.}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{output_format}{Output format of the SIA dataset.}
}
\value{
\code{NULL} silently.
}
\description{
s3_sia_cluster_dates() writes out the output of previously cached
clustered data into the local cache. This functions reads in all cached
clustered data and writes out the final SIA dataset.
}
\keyword{internal}
