% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_merge_cluster_dates_final_data}
\alias{s3_sia_merge_cluster_dates_final_data}
\title{Merge and add in all cached cluster data}
\usage{
s3_sia_merge_cluster_dates_final_data(
  sia.clean.01,
  polis_folder = Sys.getenv("POLIS_DATA_FOLDER"),
  polis_data_folder = file.path(polis_folder, "data")
)
}
\arguments{
\item{sia.clean.01}{\code{tibble} all cleaned and historical SIA data without rounds}

\item{polis_folder}{\code{str} Path to the POLIS folder.}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}
}
\value{
\code{NULL} silently.
}
\description{
s3_sia_cluster_dates() writes out the output of previously cached
clustered data into the local cache. This functions reads in all cached
clustered data and merged it into the existing SIA data
}
\keyword{internal}
