% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_es_check_metadata}
\alias{s4_es_check_metadata}
\title{Compare ES outputs with metadata from previous output}
\usage{
s4_es_check_metadata(es.05)
}
\arguments{
\item{es.05}{\code{tibble} The latest ES download with variables checked
against the last download, variables validated and sites checked and
CDC variables enforced}
}
\value{
\code{NULL} invisible return with write out to logs if necessary
}
\description{
Compare ES outputs with metadata from previous output
}
\keyword{internal}
