% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_es_create_cdc_vars}
\alias{s4_es_create_cdc_vars}
\title{Convert variables in POLIS download to to cleaned outputs}
\usage{
s4_es_create_cdc_vars(es.02, polis_folder, output_folder_name)
}
\arguments{
\item{es.02}{\code{tibble} The latest ES download with variables checked
against the last download, variables validated and sites checked}

\item{polis_folder}{\code{str} Path to the main POLIS folder.}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_\link{REGION}" (e.g., "Core_Ready_Files_AFRO").}
}
\value{
\code{tibble} es.05 SIA data with CDC variables enforced
}
\description{
Convert variables in POLIS download to to cleaned outputs
}
\keyword{internal}
