% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_es_load_data}
\alias{s4_es_load_data}
\title{Load and check ES data against previous downloads}
\usage{
s4_es_load_data(polis_data_folder, latest_folder_in_archive)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive.}
}
\value{
\code{tibble} es.01.new - the latest SIA data quality checked for variable
stability against the last download if it exists
}
\description{
Load and check ES data against previous downloads
}
\keyword{internal}
