% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_es_validate_sites}
\alias{s4_es_validate_sites}
\title{Validate ES Site names}
\usage{
s4_es_validate_sites(es.02)
}
\arguments{
\item{es.02}{\code{tibble} Tibble containing all ES data with variable contents
validated}
}
\value{
\code{tibble} es.02 - the latest SIA data quality checked for variable
stability against the last download if it exists, data cleaned and site
names validated
}
\description{
Validate ES Site names
}
\keyword{internal}
