% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_create_cdc_vars}
\alias{s5_pos_create_cdc_vars}
\title{Create CDC variables}
\usage{
s5_pos_create_cdc_vars(virus.raw.new, polis_folder, polis_data_folder)
}
\arguments{
\item{virus.raw.new}{\code{tibble} Output of \code{\link[=s5_pos_load_data]{s5_pos_load_data()}}.}

\item{polis_folder}{\code{str} Path to the main POLIS folder.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}
}
\value{
\code{tibble} Raw virus data with additional CDC columns.
}
\description{
Create variables used in CDC data analysis.
}
\keyword{internal}
