% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_create_final_virus_data}
\alias{s5_pos_create_final_virus_data}
\title{Combine cleaned human and environmental samples positives dataset}
\usage{
s5_pos_create_final_virus_data(human.virus.05, env.virus.04)
}
\arguments{
\item{human.virus.05}{\code{tibble} Output of \code{\link[=s5_pos_process_human_virus]{s5_pos_process_human_virus()}}.}

\item{env.virus.04}{\code{tibble} Output of \code{\link[=s5_pos_process_es_virus]{s5_pos_process_es_virus()}}.}
}
\value{
\code{tibble} Combined cleaned positives dataset.
}
\description{
Combines both the cleaned human and environmental surveillance samples positives
datasets into a final cleaned positives dataset.
}
\keyword{internal}
