% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_load_data}
\alias{s5_pos_load_data}
\title{Loads the most recent positives dataset}
\usage{
s5_pos_load_data(
  polis_data_folder,
  latest_folder_in_archive,
  output_folder_name
)
}
\arguments{
\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive, if one exists.}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
<<<<<<< HEAD
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").

validation.}
=======
"Core_Ready_Files_\link{REGION}" (e.g., "Core_Ready_Files_AFRO").}
>>>>>>> 4154afb (Add option to use qs file formats)
}
\value{
\code{tibble} New virus dataset.
}
\description{
Loads the most recent positives dataset and archives the previous
positives dataset.
}
\keyword{internal}
