% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_process_human_virus}
\alias{s5_pos_process_human_virus}
\title{Process and clean cases in the positives dataset}
\usage{
s5_pos_process_human_virus(
  virus.01,
  polis_data_folder,
  output_folder_name,
  output_format
)
}
\arguments{
\item{virus.01}{\code{tibble} Output of \code{\link[=s5_pos_create_cdc_vars]{s5_pos_create_cdc_vars()}}.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{output_format}{\code{str} output_format to save files as.
Available formats include 'rds' 'rda' 'csv' and 'parquet', Defaults is
'rds'.}
}
\value{
\code{tibble} Cleaned positives dataset containing human cases only.
}
\description{
Performs deep cleaning of AFP/non-AFP cases in the positives dataset.
}
\keyword{internal}
