% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_polis_key}
\alias{test_polis_key}
\title{Test out if POLIS key is valid}
\usage{
test_polis_key(key)
}
\arguments{
\item{key}{\code{str} POLIS API Key}
}
\value{
boolean
}
\description{
Test POLIS API Key
}
