% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_polis_api_call_log}
\alias{update_polis_api_call_log}
\title{Create/Update POLIS debugging API call log}
\usage{
update_polis_api_call_log(
  log_file = Sys.getenv("POLIS_LOG_FILE"),
  .time = Sys.time(),
  .call,
  .event
)
}
\arguments{
\item{log_file}{str: location of cache file}

\item{.time}{dttm: time of update}

\item{.call}{str: URL of API call conducted}

\item{.event}{str: event to be logged}
}
\value{
Return true if cache updated
}
\description{
Create/Update POLIS debugging API call log
}
