% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_polis_log}
\alias{update_polis_log}
\title{Update local POLIS interaction log}
\usage{
update_polis_log(
  log_file = Sys.getenv("POLIS_LOG_FILE"),
  .time = Sys.time(),
  .user = Sys.getenv("USERNAME"),
  .event
)
}
\arguments{
\item{log_file}{str: location of cache file}

\item{.time}{dttm: time of update}

\item{.user}{double: user who conducted the action}

\item{.event}{str: event to be logged}
}
\value{
Return true if cache updated
}
\description{
Update the POLIS log
}
