% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_gpei_cases}
\alias{add_gpei_cases}
\title{add GPEI website reported cases}
\usage{
add_gpei_cases(
  azcontainer = suppressMessages(get_azure_storage_connection()),
  proxy_data_loc = "/Data/proxy/polio_proxy_data.csv",
  polis_pos_loc = "/Data/polis/positives_2001-01-01_2025-01-06.rds"
)
}
\arguments{
\item{azcontainer}{Azure validated container object.}

\item{proxy_data_loc}{str location of proxy_data on EDAV}

\item{polis_pos_loc}{str location of latest positives dataset generated from POLIS API data}
}
\description{
a function to add manually extracted GPEI cases to positives file and estimate points
based on lowest level admin data
}
