% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_single_url.R
\name{call_single_url}
\alias{call_single_url}
\title{Call single URL}
\usage{
call_single_url(
  url,
  api_key = Sys.getenv("POLIS_API_KEY"),
  times = 10,
  diagnostics = FALSE
)
}
\arguments{
\item{url}{\code{str} Single url.}

\item{api_key}{\code{str} validated API key.}

\item{times}{\code{int} Number of times to attempt connection with API.}

\item{diagnostics}{\code{logical} Logging messages to see if calls are happening.}
}
\value{
\code{tibble} Data from the response.
}
\description{
Call a return the formatted output from one URL.
}
\details{
The function is designed for API endpoints that support skiptokens argument.
}
\examples{
\dontrun{
call_single_url("https://extranet.who.int/polis/api/v2/Virus?$filter=VirusDate\%20gt\%202025-01-01")
}

}
