% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_single_url}
\alias{call_single_url}
\title{Call single URL}
\usage{
call_single_url(url, api_key = Sys.getenv("POLIS_API_KEY"), times = 10)
}
\arguments{
\item{url}{\code{str} single url}

\item{api_key}{\code{str} validated API key}

\item{times}{\code{int} number of times to attempt connection with API}
}
\value{
tibble
}
\description{
Call a return the formatted output frome one URL
}
