% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_urls_in_parallel.R
\name{call_urls_in_parallel}
\alias{call_urls_in_parallel}
\title{Requests data from the POLIS API in parallel}
\usage{
call_urls_in_parallel(
  urls,
  requests_per_minute = 30,
  polis_key = Sys.getenv("POLIS_API_KEY")
)
}
\arguments{
\item{urls}{\code{str} A URL string or a vector of URL strings to call to the POLIS API.}

\item{requests_per_minute}{\code{int} Maximum number of requests per minute. Defaults to 30.}

\item{polis_key}{\code{str} POLIS API key.}
}
\value{
\code{tibble} Data requested from the POLIS API.
}
\description{
Requests data from the POLIS API.
}
\examples{
\dontrun{
   init_tidypolis() # initializes saved POLIS API key in the global environment
   urls <- c("https://extranet.who.int/polis/api/v2/Virus?$filter=UpdatedDate\%20ge\%202026-01-20T11:56:54.220Z\%20and\%20UpdatedDate\%20le\%202026-01-27T11:56:54.220Z",
             "https://extranet.who.int/polis/api/v2/Virus?$filter=UpdatedDate\%20gt\%202026-01-27T11:56:54.220Z")
   virus_data <- call_urls_in_parallel(urls)
}
}
