% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_new_viruses.R
\name{check_new_viruses}
\alias{check_new_viruses}
\title{Analyze errors in the Positives dataset pulled directly from the POLIS API}
\usage{
check_new_viruses(
  virus_table_path = "GID/PEB/SIR/POLIS/data/virus.rds",
  week_floor_date = lubridate::floor_date(Sys.Date(), unit = "week", week_start = 1),
  edav = TRUE
)
}
\arguments{
\item{virus_table_path}{\code{str} Absolute path to the positives file.}

\item{week_floor_date}{\code{date} Start date of the week to analyze. Defaults to the Monday of the previous week.}

\item{edav}{\code{logical} Whether to load the positives file from EDAV or locally. Defaults to TRUE.}
}
\value{
\code{list} A list containing details of the data errors, invisibly.
}
\description{
Checks for errors in the virus table downloaded from the POLIS API. It is meant
to catch errors related to inconsistencies in detections reported in the current
week.
}
\examples{
\dontrun{
check_new_viruses()
}
}
