% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_activity_table}
\alias{clean_activity_table}
\title{Clean the Activity table}
\usage{
clean_activity_table(path, crosswalk, edav = Sys.getenv("POLIS_EDAV_FLAG"))
}
\arguments{
\item{path}{\code{str} File path to the table.}

\item{crosswalk}{\code{tibble} The crosswalk table. This is the output of
\code{\link[=get_crosswalk_data]{get_crosswalk_data()}}.}

\item{edav}{\code{bool} Is the case data located on EDAV? Defaults to the
EDAV flag.}
}
\value{
\code{tibble} A cleaned Activity table.
}
\description{
The function performs cleaning of the Activity table. In particular,
de-duplication, renaming of variables via crosswalk, and removal of
empty columns.
}
\keyword{internal}
