% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_dates}
\alias{cluster_dates}
\title{A general function cluster sets of dates}
\usage{
cluster_dates(x, seed = 1234, method = "kmeans", grouping_days = 365)
}
\arguments{
\item{x}{\code{tibble} dataset with dates to be clustered. Requires that the tibble
has as a minimum a column named "sub.activity.start.date"}

\item{seed}{\code{int} defaults to 1234, used to ensure the clusters generated by
different individuals are the same}

\item{method}{\code{str} cluster method to use, can be "kmeans" (default) or "mindate"}

\item{grouping_days}{\code{int} defaults to 365. If there are not enough observations
to use the "k-means" method we will group by events taking place within a certain
period of time of each other. In this case the grouping days will be 365 or a year}
}
\value{
\code{tibble} with clusters groups and cluster methodology used
}
\description{
This function takes in a set of dates (primarily used for immunization
campaigns) and clusters them into likely "rounds" which is not readily
captured in the data using K-means clustering
}
\keyword{internal}
