% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_response_vars}
\alias{create_response_vars}
\title{This function creates a summary of SIA responses to cVDPV detections in the positives dataset}
\usage{
create_response_vars(pos)
}
\arguments{
\item{pos}{tibble a df of positive viruses}
}
\description{
a function to create summary SIA response variables for cVDPVs
}
