% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_urls.R
\name{create_table_urls}
\alias{create_table_urls}
\title{Create table URLs}
\usage{
create_table_urls(api_url, table_data, days_intervals = 7)
}
\arguments{
\item{api_url}{\code{str} Base url to be queried.}

\item{table_data}{\code{tibble} One row tibble with the data for a specific table.}

\item{days_intervals}{\code{int} Number of days for each interval. If set to \code{NULL}, then
the request will not be chunked.}
}
\value{
\code{str} Array of URLs to be used in \code{\link[=call_urls_in_parallel]{call_urls_in_parallel()}}.
}
\description{
Create URLs for a table to assist in parallelization. Each table will need to be
segmented based on their structure.
}
\details{
Valid values for \code{.table} can be found by looking at the cache via \code{\link[=get_polis_cache]{get_polis_cache()}}.
}
\keyword{internal}
