% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_urls.R
\name{create_table_urls}
\alias{create_table_urls}
\title{Create table URLs}
\usage{
create_table_urls(url, table_size, type)
}
\arguments{
\item{url}{\code{str} base url to be queried}

\item{table_size}{\code{int} integer of download}

\item{type}{\code{str} "full" or "partial"}
}
\value{
array of urls
}
\description{
create urls from table size and base url
}
