% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_table_urls}
\alias{create_table_urls}
\title{Create table URLs}
\usage{
create_table_urls(url, table_size, type)
}
\arguments{
\item{url}{str: base url to be queried}

\item{table_size}{int: integer of download}

\item{type}{str: "full" or "partial"}
}
\value{
array of urls
}
\description{
create urls from table size and base url
}
