% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.download.compare.02}
\alias{f.download.compare.02}
\title{THIS FUNCTION WOULD LIST OUT THE DISTINCT VALUES BY VARIABLE
ALL VALUES FOR NEW VARIABLE AND NEW VALUES FOR EXISTING}
\usage{
f.download.compare.02(
  df.from.f.download.compare.01,
  old.download,
  new.download,
  type = NULL
)
}
\arguments{
\item{df.from.f.download.compare.01}{\code{tibble} output from f.download.compare.01}

\item{old.download}{tibble}

\item{new.download}{tibble}

\item{type}{str name of the df that the comparison is being made on, for use in log
"AFP", "SIA", "ES", "POS"}
}
\value{
tibble of variables to compare
}
\description{
List out distinct values that are not the same by variable
}
\keyword{internal}
