% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{fetch_missing_records}
\alias{fetch_missing_records}
\title{Get extract with missed records}
\usage{
fetch_missing_records(
  table_data,
  ids = get_table_ids(table_data),
  parallel_calls = FALSE,
  old_cache = NULL,
  chunk_size = 20
)
}
\arguments{
\item{table_data}{\code{tibble} One row tibble with the data for a specific table.}

\item{ids}{\code{str} Full IDs of a table. Defaults to \code{\link[=get_table_ids]{get_table_ids()}}.}

\item{parallel_calls}{\code{logical} Whether to obtain data in parallel, or sequentially. Defaults to \code{TRUE}.}

\item{old_cache}{\code{tibble} Data associated with the table, should already by de-duplicated.}

\item{chunk_size}{\code{int} Max number of records to call per call. Using too big of a chunk
can cause issues with the API as the API call has a character length. Recommend not to
change.}
}
\value{
\verb{parquet connection} Data associated with the table including missing records.
}
\description{
Attempts to download an extract of records missed by a previous run of
\code{\link[=get_table_data]{get_table_data()}}. This will then get stored to the raw_extract folder for the
particular table.
}
