% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_crosswalk_data}
\alias{get_crosswalk_data}
\title{Get crosswalk data}
\usage{
get_crosswalk_data(
  file_loc =
    "//cdc.gov/project/CGH_GID_Active/PEB/SIR/DATA/Core 2.0/preprocessing/GetPOLIS/api_web_core_crosswalk.xlsx"
)
}
\arguments{
\item{file_loc}{str: location of crosswalk file}
}
\value{
tibble: crosswalk data
}
\description{
Get all data from crosswalk location
}
