% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_polis_cache}
\alias{get_polis_cache}
\title{Load local POLIS cache}
\usage{
get_polis_cache(cache_file = Sys.getenv("POLIS_CACHE_FILE"), .table)
}
\arguments{
\item{cache_file}{str: location of cache file}

\item{.table}{str: table to be loaded}
}
\value{
Return tibble with table information
}
\description{
Pull cache data for a particular table
}
