% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{get_polis_data}
\alias{get_polis_data}
\title{Manager function to get and update POLIS data}
\usage{
get_polis_data(type = "all", output_format = "rds")
}
\arguments{
\item{type}{choose to download population data ("pop") or all other data. Default's to "all"}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}
}
\description{
This function iterates through all tables and loads POLIS data. It
checks to ensure that new rows are created, data are updated accordingly and
deleted rows are reflected in the local system.
}
\examples{
\dontrun{
get_polis_data() #must be run after using init_tidypolis and providing a valid API key
}
}
