% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_data}
\alias{get_table_data}
\title{Request data from single table}
\usage{
get_table_data(
  .table,
  api_key = Sys.getenv("POLIS_API_KEY"),
  parallel_calls = TRUE,
  output_format = "parquet"
)
}
\arguments{
\item{api_key}{API Key}

\item{api_key}{\code{str} API Key. Defaults to the value of the global env variable POLIS_API_KEY.}

\item{parallel_calls}{\code{logical} Whether to obtain data in parallel, or sequentially. Defaults to \code{TRUE}.}

\item{output_format}{\code{str} Output format of the table. Defaults to parquet.}
}
\value{
Tibble with reference data
}
\description{
Get POLIS table Data
}
\examples{
\dontrun{
get_table_data(.table = "case")
get_table_data(.table = "virus") # must run init_tidypolis first in order to specify API key
}
}
