% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_data}
\alias{get_table_data}
\title{Request data from single table}
\usage{
get_table_data(
  api_key = Sys.getenv("POLIS_API_Key"),
  .table,
  output_format = "rds"
)
}
\arguments{
\item{api_key}{API Key}

\item{.table}{Table value to retrieve}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}
}
\value{
Tibble with reference data
}
\description{
Get POLIS table Data
}
\examples{
\dontrun{
get_table_data(.table = "case")
get_table_data(.table = "virus") #must run init_tidypolis first in order to specify API key
}
}
