% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_ids}
\alias{get_table_ids}
\title{Get Ids}
\usage{
get_table_ids(.table, .id, api_key = Sys.getenv("POLIS_API_KEY"))
}
\arguments{
\item{.table}{str: table}

\item{.id}{str: id variable}

\item{api_key}{str: POLIS API Key}
}
\value{
character array of ids
}
\description{
return Ids availalbe in table
}
