% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_size}
\alias{get_table_size}
\title{Get table size from POLIS}
\usage{
get_table_size(
  .table,
  api_key = Sys.getenv("POLIS_API_KEY"),
  cache_file = Sys.getenv("POLIS_CACHE_FILE"),
  extra_filter = ""
)
}
\arguments{
\item{.table}{str: Table to be downloaded}

\item{api_key}{str: API Key}

\item{cache_file}{str: Cache file location}

\item{extra_filter}{str: additional filtering parameters}
}
\description{
Get table size from POLIS
}
