% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_size}
\alias{get_table_size}
\title{Get table size from POLIS}
\usage{
get_table_size(
  .table,
  api_key = Sys.getenv("POLIS_API_KEY"),
  cache_file = Sys.getenv("POLIS_CACHE_FILE"),
  extra_filter = ""
)
}
\arguments{
\item{.table}{\code{str} Table to be downloaded}

\item{api_key}{\code{str} API Key}

\item{cache_file}{\code{str} Cache file location}

\item{extra_filter}{\code{str} additional filtering parameters}
}
\description{
Get table size from POLIS
}
