% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_tidypolis}
\alias{init_tidypolis}
\title{A function to intitialize a POLIS data folder}
\usage{
init_tidypolis(polis_data_folder, edav = F)
}
\arguments{
\item{polis_data_folder}{str: location of folder where to store all information}

\item{edav}{boolean: should the system use EDAV as it's cache; default FALSE}
}
\value{
Messages on process
}
\description{
Initialize API Key and local data cache for tidypolis. Inspired by the
tidycensus process for managing their API secrets.
}
\examples{
\dontrun{
init_tidypolis(polis_data_folder = "POLIS", edav = T) #create a polis data cache on CDC EDAV
init_tidypolis("C:/Users/"user_name"/Desktop/polis", edav = F)
#create a polis data cache on your local desktop
}
}
