% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_tidypolis}
\alias{init_tidypolis}
\title{A function to intitialize a POLIS data folder}
\usage{
init_tidypolis(polis_data_folder)
}
\arguments{
\item{polis_data_folder}{str: location of folder where to store all information}
}
\value{
Messages on process
}
\description{
Initialize API Key and local data cache for tidypolis. inspiration from
tidycensus process for managing their
}
