% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_cdc}
\alias{preprocess_cdc}
\title{Preprocess data process data in the CDC style}
\usage{
preprocess_cdc(polis_folder = Sys.getenv("POLIS_DATA_FOLDER"))
}
\arguments{
\item{polis_folder}{str: location of the POLIS data folder}
}
\value{
Outputs intermediary core ready files
}
\description{
Process POLIS data into analytic datasets needed for CDC
}
\keyword{internal}
