% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_cdc}
\alias{preprocess_cdc}
\title{Preprocess data process data in the CDC style}
\usage{
preprocess_cdc(polis_data_folder = Sys.getenv("POLIS_DATA_CACHE"))
}
\arguments{
\item{polis_data_folder}{str: location of polis data folder}
}
\description{
Process POLIS data into analytic datasets needed for CDC
}
