% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_cdc}
\alias{preprocess_cdc}
\title{Preprocess data process data in the CDC style}
\usage{
preprocess_cdc(
  polis_folder = Sys.getenv("POLIS_DATA_FOLDER"),
  who_region = NULL,
  output_format = "rds"
)
}
\arguments{
\item{polis_folder}{str: location of the POLIS data folder}

\item{who_region}{str: optional WHO region to filter data
Available inputs include AFRO, AMRO, EMRO, EURO, SEARO and  WPRO.}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' and 'parquet', Defaults is
'rds'.}
}
\value{
Outputs intermediary core ready files
}
\description{
Process POLIS data into analytic datasets needed for CDC
}
\keyword{internal}
