% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data retrieved from POLIS}
\usage{
preprocess_data(
  polis_folder = Sys.getenv("POLIS_DATA_FOLDER"),
  type = "cdc",
  who_region = NULL,
  output_format = "rds",
  raw_output_format = "rds",
  archive = TRUE,
  keep_n_archives = Inf
)
}
\arguments{
\item{polis_folder}{str: location of the POLIS data folder}

\item{type}{str: specify the type of preprocessing to complete}

\item{who_region}{str: optional WHO region to filter data
Available inputs include AFRO, AMRO, EMRO, EURO, SEARO and  WPRO.}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}

\item{raw_output_format}{str: raw_output_format of raw polis data based output_format
supplied to get_table_data and files saved as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}

\item{archive}{Logical. Whether to archive previous output directories
before overwriting. Default is \code{TRUE}.}

\item{keep_n_archives}{Numeric. Number of archive folders to retain.
Defaults to \code{Inf}, which keeps all archives. Set to a finite number
(e.g., 3) to automatically delete older archives beyond the N most recent.}
}
\value{
Analytic rds files
}
\description{
Create standard analytic datasets from raw POLIS data
}
\examples{
\dontrun{
preprocess_data(type = "cdc") #must run init_tidypolis to specify POLIS data location first
}
}
