% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data retrieved from POLIS}
\usage{
preprocess_data(type)
}
\arguments{
\item{type}{str: specify the type of preprocessing to complete}
}
\value{
Analytic rds files
}
\description{
Create standard analytic datasets from raw POLIS data
}
\examples{
\dontrun{
preprocess_data(type = "cdc") #must run init_tidypolis to specify POLIS data location first
}
}
