% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_spatial}
\alias{process_spatial}
\title{A function to process WHO spatial datasets}
\usage{
process_spatial(gdb_folder, output_folder, edav)
}
\arguments{
\item{gdb_folder}{str the folder location of spatial datasets, should end with .gdb,
if on edav the gdb will need to be zipped, ensure that the gdb and the zipped file name are the same}

\item{output_folder}{str folder location to write outputs to}

\item{edav}{boolean T or F, whether gdb is on EDAV or local}
}
\description{
a function to process WHO spatial datasets
}
