% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_character_dates}
\alias{remove_character_dates}
\title{function to remove original character formatted date vars from data tables}
\usage{
remove_character_dates(
  type,
  df,
  polis_data_folder = Sys.getenv("POLIS_DATA_CACHE"),
  output_folder_name
)
}
\arguments{
\item{type}{str: the table on which to remove original date vars, "AFP", "ES", "POS"}

\item{df}{tibble: the dataframe from which to remove character formatted dates}

\item{polis_data_folder}{str:  location of user's polis data folder}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_\link{REGION}" (e.g., "Core_Ready_Files_AFRO").}
}
\value{
outputs a saved reference table of original date vars and a smaller
core ready file without character dates
}
\description{
remove original date variables from POLIS tables
}
