% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_via_crosswalk}
\alias{rename_via_crosswalk}
\title{Rename variables via crosswalk}
\usage{
rename_via_crosswalk(api_data, crosswalk, table_name)
}
\arguments{
\item{api_data}{\code{tibble} data pulled from api}

\item{crosswalk}{\code{tibble} crosswalk file loaded through rio}

\item{table_name}{\code{str} name of table to be crosswalked}
}
\value{
\code{tibble} renamed table
}
\description{
Rename variables in tibble using crosswalk data
}
\keyword{internal}
