% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_single_table_diagnostic}
\alias{run_single_table_diagnostic}
\title{Run single table diagnostic}
\usage{
run_single_table_diagnostic(.table, key = Sys.getenv("POLIS_API_Key"))
}
\arguments{
\item{.table}{str: table name}

\item{key}{str: POLIS API Key}
}
\value{
tibble with diagnostic data
}
\description{
Run single table diagnostic
}
