% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_archive_old_files}
\alias{s1_archive_old_files}
\title{Archives the old files in the data folder}
\usage{
s1_archive_old_files(polis_data_folder, timestamp)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder}

\item{timestamp}{\code{str} Time stamp folder name}
}
\description{
Archives the old files in the data folder
}
\keyword{internal}
