% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_create_change_log}
\alias{s1_create_change_log}
\title{Create the change log for table data}
\usage{
s1_create_change_log(
  polis_data_folder,
  file,
  timestamp,
  api_case_data,
  api_es_data,
  api_virus_data,
  api_subactivity_data,
  output_folder_name,
  archive
)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder}

\item{file}{\code{str} Name of the file}

\item{timestamp}{\code{str} Name of the timestamp folder}

\item{api_case_data}{\code{tibble} Cleaned case data}

\item{api_es_data}{\code{tibble} Cleaned ES data}

\item{api_virus_data}{\code{tibble} Cleaned virus data}

\item{api_subactivity_data}{\code{tibble} Cleaned subactivity data}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{archive}{Logical. Whether to archive previous output directories
before overwriting. Default is \code{TRUE}.}
}
\value{
\code{NULL}, if successful
}
\description{
Create the change log for table data
}
\keyword{internal}
