% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_export_final_core_ready_files}
\alias{s1_export_final_core_ready_files}
\title{Export the final Core Ready files}
\usage{
s1_export_final_core_ready_files(
  polis_data_folder,
  ts,
  timestamp,
  api_case_data,
  api_subactivity_data,
  api_es_data,
  api_virus_data,
  output_folder_name,
  output_format
)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{ts}{\code{ts} A Sys.time() output.}

\item{timestamp}{\code{str} Formatted time stamp.}

\item{api_case_data}{\code{tibble} Cleaned Case table.}

\item{api_subactivity_data}{\code{tibble} Cleaned Subactivity table.}

\item{api_es_data}{\code{tibble} Cleaned ES table.}

\item{api_virus_data}{\code{tibble} Cleaned Virus table.}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}
}
\value{
\code{NULL}
}
\description{
Export the final Core Ready files
}
\keyword{internal}
