% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_prep_polis_tables}
\alias{s1_prep_polis_tables}
\title{Prepare the POLIS tables for preprocessing}
\usage{
s1_prep_polis_tables(
  polis_folder,
  polis_data_folder,
  long.global.dist.01,
  ts,
  timestamp,
  who_region,
  output_format,
  raw_output_format,
  archive
)
}
\arguments{
\item{polis_folder}{str: location of the POLIS data folder}

\item{polis_data_folder}{\code{str} Data folder within the POLIS folder.}

\item{long.global.dist.01}{\code{sf} Long global district shapefile.}

\item{ts}{\code{str} Time stamp from \code{\link[=Sys.time]{Sys.time()}}.}

\item{timestamp}{\code{str} Formatted time stamp.}

\item{who_region}{str: optional WHO region to filter data
Available inputs include AFRO, AMRO, EMRO, EURO, SEARO and  WPRO.}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}

\item{raw_output_format}{str: raw_output_format of raw polis data based output_format
supplied to get_table_data and files saved as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}

\item{archive}{Logical. Whether to archive previous output directories
before overwriting. Default is \code{TRUE}.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Prepares the POLIS tables downloaded via API for further preprocessing.
}
