% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s1_prep_polis_tables}
\alias{s1_prep_polis_tables}
\title{Prepare the POLIS tables for preprocessing}
\usage{
s1_prep_polis_tables(
  polis_folder,
  polis_data_folder,
  long.global.dist.01,
  ts,
  timestamp
)
}
\arguments{
\item{polis_folder}{str: location of the POLIS data folder}

\item{polis_data_folder}{\code{str} Data folder within the POLIS folder.}

\item{long.global.dist.01}{\code{sf} Long global district shapefile.}

\item{ts}{\code{str} Time stamp from \code{\link[=Sys.time]{Sys.time()}}.}

\item{timestamp}{\code{str} Formatted time stamp.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Prepares the POLIS tables downloaded via API for further preprocessing.
}
