% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s2_check_duplicated_epids}
\alias{s2_check_duplicated_epids}
\title{Check for Duplicated EPIDs in AFP Data}
\usage{
s2_check_duplicated_epids(data, polis_data_folder, output_folder_name)
}
\arguments{
\item{data}{A dataframe containing AFP surveillance data}

\item{polis_data_folder}{String path to the POLIS data folder}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}
}
\value{
Logical. TRUE if duplicates found, FALSE if no duplicates
}
\description{
This function checks for duplicate EPID entries in AFP surveillance data and
exports duplicates to a CSV file if found.
}
\details{
The function:
\enumerate{
\item Processes dates and extracts year of onset
\item Groups by EPID, onset date and admin level 0 location
\item Identifies records with 2+ entries for same EPID
\item Exports duplicates to CSV if found
\item Logs the duplicate finding in POLIS system
}
}
\examples{
\dontrun{
s2_check_duplicated_epids(afp_data, "path/to/polis/folder")
}
}
\keyword{internal}
