% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_fully_process_sia_data}
\alias{s3_fully_process_sia_data}
\title{Process SIA Data Pipeline}
\usage{
s3_fully_process_sia_data(long.global.dist.01)
}
\arguments{
\item{long.global.dist.01}{\code{sf} Global district lookup table for GUID
validation.}
}
\description{
This function processes SIA data through multiple standardization and
validation steps, including checking for duplicates, standardizing dates,
clustering dates, and processing coordinates.
}
