% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_fully_process_sia_data}
\alias{s3_fully_process_sia_data}
\title{Process SIA Data Pipeline}
\usage{
s3_fully_process_sia_data(
  long.global.dist.01,
  polis_data_folder,
  latest_folder_in_archive,
  timestamp,
  output_folder_name
)
}
\arguments{
\item{long.global.dist.01}{\code{sf} Global district lookup table for GUID
validation.}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive.}

\item{timestamp}{\code{str} The time stamp.}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_\link{REGION}" (e.g., "Core_Ready_Files_AFRO").}
}
\description{
This function processes SIA data through multiple standardization and
validation steps, including checking for duplicates, standardizing dates,
clustering dates, and processing coordinates.
}
