% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_load_data}
\alias{s3_sia_load_data}
\title{Load SIA Data}
\usage{
s3_sia_load_data(
  polis_data_folder,
  latest_folder_in_archive,
  output_folder_name
)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{latest_folder_in_archive}{\code{str} Time stamp of latest folder in archive}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}
}
\value{
\code{tiblle} sia.01.new - the latest SIA data quality checked for variable
stability against the last download if it exists
}
\description{
Load SIA Data
}
\keyword{internal}
