% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s3_sia_write_precluster_data}
\alias{s3_sia_write_precluster_data}
\title{Write out SIA data}
\usage{
s3_sia_write_precluster_data(
  sia.06,
  polis_data_folder,
  output_folder_name,
  output_format
)
}
\arguments{
\item{sia.06}{\code{tibble} The latest SIA download with variables checked and
against the last download, CDC variables created, GUIDs validated and
deduplicated}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' 'qs' and 'parquet', Defaults is
'rds'.}
}
\description{
Write out SIA data
}
\keyword{internal}
