% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_es_create_cdc_vars}
\alias{s4_es_create_cdc_vars}
\title{Convert variables in POLIS download to to cleaned outputs}
\usage{
s4_es_create_cdc_vars(es.02, polis_folder)
}
\arguments{
\item{es.02}{\code{tibble} The latest ES download with variables checked
against the last download, variables validated and sites checked}

\item{polis_folder}{str: location of the POLIS data folder}
}
\value{
\code{tibble} es.05 SIA data with CDC variables enforced
}
\description{
Convert variables in POLIS download to to cleaned outputs
}
\keyword{internal}
