% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_fully_process_es_data}
\alias{s4_fully_process_es_data}
\title{Process ES Data Pipeline}
\usage{
s4_fully_process_es_data(
  polis_folder,
  polis_data_folder = file.path(polis_folder, "data"),
  latest_folder_in_archive,
  output_format
)
}
\arguments{
\item{polis_folder}{str: location of the POLIS data folder}

\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive.}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' and 'parquet', Defaults is
'rds'.}
}
\value{
\code{NULL} silently upon success.
}
\description{
This function processes ES data through multiple standardization and
validation steps, including checking for duplicates, validating
ES sites and checking against previous downloads
}
