% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s4_fully_process_es_data}
\alias{s4_fully_process_es_data}
\title{Process ES Data Pipeline}
\usage{
s4_fully_process_es_data(polis_data_folder, latest_folder_in_archive)
}
\arguments{
\item{polis_data_folder}{\code{str} Path to the POLIS data folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive.}
}
\description{
This function processes ES data through multiple standardization and
validation steps, including checking for duplicates, validating
ES sites and checking against previous downloads
}
