% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_fully_process_pos_data}
\alias{s5_fully_process_pos_data}
\title{Process positives data pipeline}
\usage{
s5_fully_process_pos_data(
  polis_folder,
  latest_folder_in_archive,
  long.global.dist.01,
  polis_data_folder = file.path(polis_folder, "data"),
  output_folder_name
)
}
\arguments{
\item{polis_folder}{\code{str} Path to the main POLIS folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive, if one exists.}

\item{long.global.dist.01}{\code{sf} Global district lookup table for GUID}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_\link{REGION}" (e.g., "Core_Ready_Files_AFRO").

validation.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
This function processes virus data through multiple standardization and
validation steps, including checking for duplicates, standardizing dates,
classifying cases
}
