% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_check_duplicates}
\alias{s5_pos_check_duplicates}
\title{Check for duplicate entries in the virus table}
\usage{
s5_pos_check_duplicates(virus.01, polis_data_folder, output_folder_name)
}
\arguments{
\item{virus.01}{\code{tibble} Output of \code{\link[=s5_pos_create_cdc_vars]{s5_pos_create_cdc_vars()}}.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_\link{REGION}" (e.g., "Core_Ready_Files_AFRO").

validation.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Checks for potential duplicate positive records and outputs the result to
a csv in \code{polis_data_folder}
}
\keyword{internal}
