% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_compare_with_archive}
\alias{s5_pos_compare_with_archive}
\title{Compares the old virus dataset in the latest archive folder with the new processed virus dataset}
\usage{
s5_pos_compare_with_archive(
  afp.es.virus.01,
  afp.es.virus.03,
  polis_data_folder,
  latest_folder_in_archive,
  output_folder_name,
  output_format
)
}
\arguments{
\item{afp.es.virus.01}{\code{tibble} Output of \code{\link[=s5_pos_create_final_virus]{s5_pos_create_final_virus()}}.}

\item{afp.es.virus.03}{\code{tibble} Output of \code{\link[=s5_pos_create_final_virus]{s5_pos_create_final_virus()}} but processed
further using \code{\link[=remove_character_dates]{remove_character_dates()}} and \code{\link[=create_response_vars]{create_response_vars()}}.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive, if one exists.}

\item{output_folder_name}{str: Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").

validation.}

\item{output_format}{str: output_format to save files as.
Available formats include 'rds' 'rda' 'csv' and 'parquet', Defaults is
'rds'.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Compares the old virus dataset with the new processed virus dataset to determine
any changes, including the number of records that changed since the last run. Changes
are also added into the \code{polis_data_folder} as csv files.
}
\keyword{internal}
