% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_load_data}
\alias{s5_pos_load_data}
\title{Loads the most recent positives dataset}
\usage{
s5_pos_load_data(polis_data_folder, latest_folder_in_archive)
}
\arguments{
\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.
validation.}

\item{latest_folder_in_archive}{\code{str} Name of the latest folder in the archive, if one exists.}
}
\value{
\code{tibble} New virus dataset.
}
\description{
Loads the most recent positives dataset and archives the previous
positives dataset.
}
\keyword{internal}
