% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_process_es_virus}
\alias{s5_pos_process_es_virus}
\title{Process and clean cases in the positives dataset}
\usage{
s5_pos_process_es_virus(virus.01, polis_data_folder)
}
\arguments{
\item{virus.01}{\code{tibble} Output of \code{\link[=s5_pos_create_cdc_vars]{s5_pos_create_cdc_vars()}}.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.
validation.}
}
\value{
\code{tibble} Cleaned positives dataset containing environmental samples only.
}
\description{
Performs deep cleaning of environmental samples in the positives dataset.
}
\keyword{internal}
