% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_write_missing_onsets}
\alias{s5_pos_write_missing_onsets}
\title{Obtain positives records with missing onset date}
\usage{
s5_pos_write_missing_onsets(virus.01, polis_data_folder, output_folder_name)
}
\arguments{
\item{virus.01}{\code{tibble} Output of \code{\link[=s5_pos_create_cdc_vars]{s5_pos_create_cdc_vars()}}.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.}

\item{output_folder_name}{\code{str} Name of the output directory where processed
files will be saved. Defaults to "Core_Ready_Files". For
region-specific processing, this should be set to
"Core_Ready_Files_REGION" (e.g., "Core_Ready_Files_AFRO").}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Obtains the records in the positives dataset without onset date and outputs them to the
\code{polis_data_folder} as a csv.
}
\keyword{internal}
