% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s5_pos_write_missing_onsets}
\alias{s5_pos_write_missing_onsets}
\title{Obtain positives records with missing onset date}
\usage{
s5_pos_write_missing_onsets(virus.01, polis_data_folder)
}
\arguments{
\item{virus.01}{\code{tibble} Output of \code{\link[=s5_pos_create_cdc_vars]{s5_pos_create_cdc_vars()}}.}

\item{polis_data_folder}{\code{str} The data folder within the POLIS folder.
validation.}
}
\value{
\code{NULL} quietly upon success.
}
\description{
Obtains the records in the positives dataset without onset date and outputs them to the
\code{polis_data_folder} as a csv.
}
\keyword{internal}
