% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypolis-package.R
\docType{package}
\name{tidypolis-package}
\alias{tidypolis}
\alias{tidypolis-package}
\title{tidypolis: Download and process POLIS data}
\description{
This package uses the POLIS API to ease download and processing of global Polio data.
}
\author{
\strong{Maintainer}: Nishant Kishore \email{ynm2@cdc.gov} (\href{https://orcid.org/0000-0003-0408-2747}{ORCID})

Authors:
\itemize{
  \item Nicholas Heaghney \email{uic3@cdc.gov}
  \item Mervin Keith Cuadera \email{xrg9@cdc.gov} (\href{https://orcid.org/0000-0003-4898-2659}{ORCID})
}

}
\keyword{internal}
