% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{tidypolis_io}
\alias{tidypolis_io}
\title{Tiypolis input and output}
\usage{
tidypolis_io(
  obj = NULL,
  io,
  file_path,
  edav = Sys.getenv("POLIS_EDAV_FLAG"),
  azcontainer = suppressMessages(sirfunctions::get_azure_storage_connection()),
  full_names = F
)
}
\arguments{
\item{obj}{str: object to be loaded into EDAV}

\item{io}{str: read/write/list/exists/create/delete}

\item{file_path}{str: absolute path of file}

\item{edav}{boolean defaults to FALSE}

\item{azcontainer}{AZ container object returned by}

\item{full_names}{boolean: If you want to include the full reference path in the response, dfault FALSE}
}
\value{
conditional on `io`
}
\description{
Manages read/write/list/create/delete functions for tidypolis
}
