% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_polis_cache}
\alias{update_polis_cache}
\title{Update local POLIS cache}
\usage{
update_polis_cache(
  cache_file = Sys.getenv("POLIS_CACHE_FILE"),
  .table,
  .nrow,
  .update_val
)
}
\arguments{
\item{cache_file}{str: location of cache file}

\item{.table}{str: table to be updated}

\item{.nrow}{double: nrow of table to be updated}

\item{.update_val}{str: value to update data}
}
\value{
Return true if cache updated
}
\description{
Update the POLIS cache directory
}
\keyword{internal}
