% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{update_polis_table}
\alias{update_polis_table}
\title{Update the POLIS table}
\usage{
update_polis_table(table_data, table_url, parallel_calls = TRUE, output_format)
}
\arguments{
\item{table_data}{\code{tibble} One row tibble with the data for a specific table.}

\item{table_url}{\code{str} URL of the endpoint for the specific table.}

\item{parallel_calls}{\code{logical} Whether to obtain data in parallel, or sequentially. Defaults to \code{TRUE}.}

\item{output_format}{\code{str} Output format of the table. Defaults to parquet.}
}
\value{
\code{NULL}
}
\description{
Will download data above the update date recorded in the metadata cache.
}
\keyword{internal}
