% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{upload_cdc_proc_to_edav}
\alias{upload_cdc_proc_to_edav}
\title{Transfer preprocessed files to active EDAV download}
\usage{
upload_cdc_proc_to_edav(
  core_ready_folder = file.path(Sys.getenv("POLIS_DATA_CACHE"), "Core_Ready_Files"),
  azcontainer = sirfunctions::get_azure_storage_connection(),
  output_folder = "Data/polis"
)
}
\arguments{
\item{core_ready_folder}{str: Local folder with CDC processed files}

\item{azcontainer}{Azure Token Container Object}

\item{output_folder}{Location to write out Core Files}
}
\description{
Manages read/write/list/create/delete functions for tidypolis
}
