% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{upload_cdc_proc_to_edav}
\alias{upload_cdc_proc_to_edav}
\title{Transfer preprocessed files to active EDAV download}
\usage{
upload_cdc_proc_to_edav(
  core_ready_folder = file.path(Sys.getenv("POLIS_DATA_CACHE"), "Core_Ready_Files"),
  azcontainer = sirfunctions::get_azure_storage_connection(),
  output_folder = "Data/polis"
)
}
\arguments{
\item{core_ready_folder}{\code{str} Local folder with CDC processed files.}

\item{azcontainer}{\verb{Azure container} Azure Token Container Object.}

\item{output_folder}{\code{str} Location to write out Core Files.}
}
\value{
\code{NULL}.
}
\description{
The function will perform an archiving operation of the previous preprocessed
files in the data folder used when recreating \code{raw.data} with updated dataset.
It will also copy the newly preprocessed files to the data folder.
}
