% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_urls.R
\name{week_cuts_api}
\alias{week_cuts_api}
\title{Cuts the update dates into time intervals}
\usage{
week_cuts_api(anchor_date, update_col, days_intervals = NULL)
}
\arguments{
\item{anchor_date}{\code{str} The latest update date from a table.}

\item{update_col}{\code{str} Name of the column with the update date, if applicable.}

\item{days_intervals}{\code{int} Number of days for each interval.}
}
\value{
\code{str} A vector of strings representing date intervals.
}
\description{
Cuts the time intervals based on an anchor date, with the last record
being open ended.
}
\details{
Majority of the code in this function was generated using Microsoft 365 Copilot
Version number: 2.20260115.41.0. The code was reviewed and
edited by Mervin Cuadera (xrg9).
}
\keyword{internal}
